pyproject.toml
[project]
name = "opennet-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jsonschema>=4.24.0",
    "pytest>=8.4.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]
